Задел на ручку
Добавлены роли, доступы, 1 класс и 1 html файл для будущей собственной страницы входа

+++ src/main/java/traffic_id/demo/entity/Product.java
@@ -12,7 +12,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
 
@@ -20,16 +20,16 @@ public class Product {
     public Product() {
     }
 
-    public Product(Long id, String name) {
+    public Product(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
